Workaround hack for Trådfri
When modifying the dim value with the dedicated IKEA remote, the lights in the group and the group itself does not have the same dim value anymore. When trying to set the same value as the group already believe it has, it will not do anything. This is a hacky workaround for this, which works by internally storing if this mismatch is true for the group. If it is, we simply modify the value we set the next time by one. No one should be able to notice that small change, but it makes the code messier.